+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Table of contents
-- [Release Candidate 30 (TBD)](#release-candidate-30-tbd)
+- [Release Candidate 31 (TBD)](#release-candidate-31-tbd)
+- [Release Candidate 30 (31 Dec 2021)](#release-candidate-30-31-dec-2021)
 - [Release Candidate 29 (07 Nov 2021)](#release-candidate-29-07-nov-2021)
 - [Release Candidate 28 (06 Sep 2021)](#release-candidate-28-06-sep-2021)
 - [Release Candidate 27 (03 Sep 2021)](#release-candidate-27-03-sep-2021)
@@ -30,13 +31,25 @@
 - [Release Candidate 2 (29 Sep 2019)](#release-candidate-2-29-sep-2019)
 - [Release Candidate 1 (26 Sep 2019)](#release-candidate-1-26-sep-2019)
 
-## Release Candidate 30 (TBD)
+## Release Candidate 31 (TBD)
+
+#### Additions
+
+#### Changes
+
+#### Fixes
+* Fixed #3390
+
+## Release Candidate 30 (31 Dec 2021)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#30
 
 #### Additions
 * Added a ton of wiki links to the guide
 * (API) Added "GRAVITY_AFFECTED_BLOCKS" tag
 * (API) Added "Biome-Maps" for more in-depth GEO resource configuration (developers only for now)
+* (API) Added some utility methods for Biome-Maps
 * Added support for 1.18
+* Added Talisman of Farmer
 
 #### Changes
 * GEO resource distributions have been slightly adjusted
@@ -45,8 +58,12 @@
 #### Fixes
 * Crimson and Warped Pressure Plates are now properly recognized as pressure plates
 * Fixed #3336
+* (API) Fixed `Parachute` constructor parameter being ignored
+* Fixed #3385
+* Fixed (Easter) Apple Pie recipe yielding (Christmas) Apple Pies
 
 ## Release Candidate 29 (07 Nov 2021)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#29
 
 #### Additions
 * Added support for deepslate ores and copper with the Hercules' Pickaxe
@@ -75,12 +92,14 @@
 * Fixed smithing table issue on 1.15 and lower
 
 ## Release Candidate 28 (06 Sep 2021)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#28
 
 #### Fixes
 * Fixed Metrics
 * Fixed some naming conventions and localization keys for RC-27
 
 ## Release Candidate 27 (03 Sep 2021)
+https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#27
 
 ### **Breaking Changes (API)**
 This RC brings a lot of breaking changes to the API. For more info on why we did this and what happened [please refer to our PSA](https://github.com/Slimefun/Slimefun4/pull/3139)